% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{gbrt}
\alias{gbrt}
\title{Imitate gradient boost regression trees}
\usage{
gbrt(train_data, test_data, target, learning_rates = 0.1, n_trees = 100)
}
\arguments{
\item{train_data}{split the original data randomly into 2 datasets, one is training dataset, train data is its features}

\item{test_data}{split the original data randomly into 2 datasets, another one is testing dataset, test_data is its features}

\item{target}{dependent variable to predict}

\item{learning_rates}{it controls how much we adjust the weights of our network with respect to the loss gradient}

\item{n_trees}{number of trees}
}
\value{
prediction of the target variable
}
\description{
Imitate gradient boost regression trees
}
\examples{
library(rpart)
library(tidyr)
library(tidyverse)
library(gbm)
library(palmerpenguins)
library(caret)
data("penguins")
penguins <- na.omit(penguins)
penguins <- penguins \%>\% select(-year)
set.seed(123)
indices <- sample(1:nrow(penguins), 0.7 * nrow(penguins))
train_data <- penguins[indices, ]
test_data <- penguins[-indices, ]
target <- "body_mass_g"
predictions <- gbrt(train_data,test_data,target)
}
