% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{gbrt_cv}
\alias{gbrt_cv}
\title{use cross-validation to tune hyperparameter(learning rate)}
\usage{
gbrt_cv(data, target, learning_rates, n_trees = 100, folds = 5)
}
\arguments{
\item{data}{dataset use here}

\item{target}{dependent variable to predict}

\item{learning_rates}{it controls how much we adjust the weights of our network with respect to the loss gradient}

\item{n_trees}{number of trees}

\item{folds}{number of folds}
}
\value{
contains information about the learning rate and its corresponding rmse
}
\description{
use cross-validation to tune hyperparameter(learning rate)
}
\examples{
library(rpart)
library(tidyr)
library(tidyverse)
library(gbm)
library(palmerpenguins)
library(caret)
data("penguins")
penguins <- na.omit(penguins)
penguins <- penguins \%>\% select(-year)
learning_rates <- c(0.01, 0.05, 0.1, 0.2, 0.5)
cv_results <- gbrt_cv(penguins, target = "body_mass_g", learning_rates,n_trees = 100,folds = 5)
}
